+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Fix `:dependency-scheme` setting for .java files from jars #1653
   - Bump graalvm to 21.
   - Improve overall performance using GraalVM 21 PGO (Profile-Guided Optimizations).
+  - Extract the responsibility for merging clj-depend config #1265
 
 - Editor
   - New paredit refactorings: